refactor: rename utils to services
